Foundations for supporting methods (and more)
We want to be able to express RakuAST much like a document - that is to
say, a class has methods within it. However, these are not directly held
(and need not be directly held). Thus there's a need to have them
attached to the appropriate package node ahead of the point we go making
the meta-object. A related problem is placeholder parameters, which can
appear deep within an expression, but need attaching to their enclosing
target block.

Thus an attaching mechanism is introduced for this. It is used to add
methods. However, those methods - while existing in the method table -
are not actually yet callable, because nothing knows about the implicit
invocant parameter. Still, it's enough to get method showing up there.

Attribute accesses may well be resolved and checked through this kind of
mechanism too, because we can't actually know if they exist - or how to
compile them if they're natives - until after class composition time.
(The current compiler is broken in this regard, and can go wrong when a
natively typed attribute comes from a role.)