Handle explicitly managed strings getting passed to native functions
For explicitly managed strings, we need to pass the contained CStr value
instead of the string object itself. This way NativeCall knows not to free
the created C string after the call.

Note that the generated function bodies have never supported this. The problem
just hasn't surfaced before, because the test runs the call only once and in
the previous implementation, the first call always went through the generic
calling code (which does support explicitly managed strings).